When generating the state space of 1394-fin using the compiling rewriter two observations were made regarding equalities on functions.
There were not enough functions generated of the shape make_term_with_many_arguments.

The functions for equality on functions were wrong due to an index that was not 
increased, i.e. they had the shape:

   f == g  =   forall x0:A,x0:B,.... f(x0,x0,...)==g(x0,x0,...)

which is not quite what was intended.

A test has been added to catch this.

Furhtermore, an application can now be applied on 7 arguments also, as this is now
also possible in the aterm library.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13699 21252fe5-60e8-0310-8ff0-c1ae1001594e