Improved the linearisation of stochastic processes further. There was a missing substitution of arguments in a newly created process for a stochastic process. This has been repaired.
Furthermore, the variables of a stochastic process were not cleaned
up in a substitution, which caused the enumerator to get confused.
It appears the enumerator works fine now.

Furthermore, a puzzle which came to me via Onno Boxma has been added
where people enter a plane. The first passenger lost its boarding pass
and will sit at a random place. The question is what the probability is
that the last passenger will have its own place. 
The probabilistic state space for this can be generated (up to 
hundreds of thousands of passengers). For small numbers I can manually
verify that the remarkable answer matches that of the generated transition
system.

To calculate the answer for large numbers of passengers
a behavioural reduction tool or a probabilistic
model checker is required. Unfortunately, none of these are available yet.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13920 21252fe5-60e8-0310-8ff0-c1ae1001594e