Made the a substitution object in pbesinst_alternative_lazy_algorithm.h lightweight.
Replaced the more costly 3 valued variables in lpsrealelm by two valued variables.
In the previous version, there were variables with valued smaller, equal and larger.
In the current version these variables are booleans, that mean smaller and larger or
equal, or smaller or equal and larger. In the current setup sometimes more variables
are needed to replace a given set of inequalities of a process. So, it is not really
clear whether the new approach is always beneficial.

Moved the linear inequalities unit to the data library.

Some minor improvements. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13944 21252fe5-60e8-0310-8ff0-c1ae1001594e