Exercising with applying the allow operator with the empty set led to the conclusion that the example
init allow({},tau||tau);

did not have a proper Terminate action, 
whereas 

init allow({a},tau||tau);

did. The reason was that the check for the Terminate action was just at the wrong place, 
not being carried out when the allow set would be empty. Now things should work fine.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14122 21252fe5-60e8-0310-8ff0-c1ae1001594e