Added the flag --tau to lps2lts to explicitly indicate which actions should be considered hidden when searching for a divergence loop.
Removed a nasty bug where stochastic_action_summands where silently
converted to action_summands and back, changing their distribution.

It is not yet possible to print a counter example of the divergence.
The intention is that such a counter example will be printed using
the actions mentioned in the --tau option. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14142 21252fe5-60e8-0310-8ff0-c1ae1001594e