Added a mapping hidden_labels_map to labelled transition systems. This map indicates for each action label index which action label index belongs to it after hiding some action labels. Because we have multi-actions this map can be quite complex, i.e. action
   a|b

has as its corresponding action after hiding b the action

   a

And if a and b are both hidden, its corresponding action is 

   tau.

The need for this distinction is that when generating counter examples
these are not very insightfull when all actions are tau. Rather one
takes the visible actions and indicates which of these need to be 
seen as tau's, calculates an equivalence in terms of these hidden actions,
but provide a counterexample in terms of the visible actions.

There are still problems with generating counterexamples and hiding
when applying bisimulation.

Furthermore, the m_taus array in labelled transition systems has 
become superfluous and needs to be removed. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14145 21252fe5-60e8-0310-8ff0-c1ae1001594e