Adapted comments to the article text (in particular line numbers).
Added a small optimisation:  a list of new bottom states is only needed if the
first refinement (Algorithm 2, Line 2.21) produced a red block that has to be
split again (in Line 2.22) before the new bottom states are processed.
Otherwise, the new bottom states are only stored in a separate slice in the
block (as unmarked bottom states), where they will be handled immediately
afterwards.

Added another optimisation: Based on Anton's numbers, it seems that using the
Fibonacci numbers for balancing the work are the most efficient.  I therefore
added, in Release mode, balancing with Fibonacci numbers.  (In Debug mode,
strict balancing has to be kept because the iteration counters only allow for
a single step of disbalance.)



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14202 21252fe5-60e8-0310-8ff0-c1ae1001594e