Resolved the failing tests:
       1002 - tooltest_txt2pbes_abp_nodeadlock_pbes_txt (Failed)
       1088 - tooltest_txt2pbes_abp_nolivelock_pbes_txt (Failed)
       6205 - regression_ticket-1301 (Failed)

The reason for the problem was the following. From now on it is possible
to add action and data specifications to modal formulas. When parsing the
data and actions of the surrounding lps were added to the modal formula specification.
Subsequently, the extended data and actions were added to the linear process.
This caused data types to occur twice, leading to problems.

Also adapted some compiler warnings in SmallProgressMeasures.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14415 21252fe5-60e8-0310-8ff0-c1ae1001594e