- Use map.find instead of map.count to determine if an element is in a map. Put it in a function has_key to make the code more readable.
- Use map.at instead of map::operator[] to access a map in a read only way.

Made it very explicit what side effects the various optimizations have on the attributes of the pbesinst algorithm.

There are still a few questionable dependencies left. For example, it seems suspicious that the justification attribute is used for more purposes than printing a justification of how the result was obtained. And it looks like the reachable attribute should not be part of the main algorithm at all. And it looks like the trivial attribute is redundant.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14490 21252fe5-60e8-0310-8ff0-c1ae1001594e