Repaired a particular stupid error in mcrl2i (for which I am most likely to blame). A piece of code read:
   rebuild_rewriter = rebuild_rewriter || add_functions_data_specification(new_functions)

The purpose of this was to add the new functions to the data specification and if that would be
done set rebuild_rewriter to true.

Of course the actual effect was that only if rebuild_rewriter was false, this would happen.
If rebuild_rewriter was true, the new functions would not even be added to the data specification.

This has been resolved now, dealing with bug_report #1380.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14529 21252fe5-60e8-0310-8ff0-c1ae1001594e