Simplified the data type checker.
The functions

TraverseVarConsTypeD
TraverseVarConsTypeDN
UpCastNumericType

had as arguments AllowedVars and FreeVars.
These turned out to be used only to check that 
the variable in the right hand side and the 
condition of an equation also occurred in the
right hand side. 

This check has been isolated allowing to remove
AllowedVars and FreeVars. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14572 21252fe5-60e8-0310-8ff0-c1ae1001594e