mcrl2xi: some visual changes, some bug fixes
I removed floating window with only the 'parse' button
in it, and replaced it with a toolbar instead. The icons
on this toolbar only work under Linux for now.
Bugfixes: clicking cancel in the 'do you want to save
this file' dialog window crashed the tool; some keyboard
shortcuts were not working.

Fixes #1382


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14578 21252fe5-60e8-0310-8ff0-c1ae1001594e