This commit deals with error #1389 indicating that with the -DNDEBUG -O0 compilation would lead to linking errors, where errors would not occur with -DNDEBUG -O2 or with -O2 without the debug flag.
The problem was caused because the the required objects were only created in .cpp files
when NDEBUG was not defined. If NDEBUG was defined the objects were not accessed when
the -O2 flag was used. In that case the objects were translated away. With the -O0 flag the
compiler did need the objects. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14583 21252fe5-60e8-0310-8ff0-c1ae1001594e