Repaired an appalling error where the reference count of function symbols was insufficiently increased. This applied only to function symbols that were created multiple times with the same name and arity.
Oddly enough only the Ubuntu tests appeared to fail on this, while runs on other
platforms succeeded.

Added additional internal asserts to check that function_symbols are always valid
when used.

Also added a static_assert(false,..) to check whether some pieces of WIN32 code is still
being executed. I expect it is not. This allows to remove these pieces of code.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14630 21252fe5-60e8-0310-8ff0-c1ae1001594e