On Ubuntu and Fedora it appears that static function_symbols are destroyed after the function_symbol_store is destroyed. Added a check when destroying a function symbol on whether the function symbol store still exists. If not, destruction (meaning decreasing its reference count, and removing it from the function symbol store if the reference count is zero), is not performed.
Problems of a similar nature, regarding unpredictable destruction
of statically defined variables in C++, did haunt us for quite some
time with the previous implementation of function symbols and the 
current implementation of aterms. 

It appears that on Fedora all tests succeed again, except for
librarytest_mcrl2_data_quantifier_test.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14633 21252fe5-60e8-0310-8ff0-c1ae1001594e