Saving .lts files has become much more efficient. The disk footprint (including state labels) is less than that of a .aut file. The memory footprint for an .lts is ten times worse and it takes appr. twice the time to generate a .lts, where most of the time goes into saving.
However large files can now be saved in .lts format. Unfortunately,
the stack is not sufficiently big to deal with the term_lists occurring in
the .lts format. Therefore, all standard list handling routines have
been modified to use the stack for short lists (up to 10000) elements,
and the heap if the lists are longer. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14649 21252fe5-60e8-0310-8ff0-c1ae1001594e