fix opengl header test
openglwrapper.h is now implemented using macros, this way it needs much less
boilerplate code than with wrapper functions. in the best case they are just
temporary workarounds until i figure out a better way to do it (with Qt or
    portable OpenGL extensions)

furthermore, replace the deprecated gluBuild2DMipmaps with glGenerateMipmap.
the latter is technically an OpenGL extension, but it seems to be the best
fit. it should be present on all opengl implementations >= 3.0


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14732 21252fe5-60e8-0310-8ff0-c1ae1001594e