While studying why some examples were dreadfully slow, it was detected that function symbols started with a reference_count of 1 instead of 0, which led to the situation that they were never removed anymore.
The effect of this was that the generation of fresh variables became very
slow as each new fresh variable is a new function symbol. This slowed the
enumerator down, as well as rewriting where, lambda's, sum operators and
quantifications. In some cases this could slow down the generation of 
a state space with factors over 100, while using more than 1000 times
more memory. 

Replacing the crucial 1 with 0 resolved this issue. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14778 21252fe5-60e8-0310-8ff0-c1ae1001594e