Disabled renaming of state numbers when reading a .aut file. When generating a .aut file using bitstatehashing state numbers could be arbitrary, representing the position of its associated bit in the bit hash table. This required renumbering of the state numbers when reading such a file.
This has now been changed by consecutively numbering the states
when using bit hashing.

From now on it is not possible to read a .aut file of which the
state indices are larger than the number of states. 

The renumbering of states turned out to be very confusing for those
who were studying the resulting transition system in for instance
ltsgraph where the state numbers would differ from those in the
.aut file. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14813 21252fe5-60e8-0310-8ff0-c1ae1001594e