lpsxsim: Some changes to the column widths
The column that holds parameter names is now resized to fit its contents.

Doing the same for columns that hold action names is very hard, since
it is not possible to predict the maximum possible length of actions
for a certain specification. Continously resizing columns automatically
is not user friendly. Therefore, I have chosen to just give them a
default width of 180.

fixes #1396


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14857 21252fe5-60e8-0310-8ff0-c1ae1001594e