Changed the way ultimate delays are being generated within the lineariser. Thomas Neele indicated that in the current lineariser the ultimate delays would blow up exponentially as they were calculated on the basis of the parallel composition.
In the new approach each mCRL process also comes with its own ultimate
delay. So for processes p1 || p2, there are ultimate delays u1 and u2.
In the parallel composition u1 is applied to the non communicating 
summands from p2, and u2 is applied to those of u1. The ultimate
delay of the parallel composition is u1 && u2.

This handles report #1425 where it is indicated that on fischer's 
protocol with five processes, or the train gate with four trains
the lineariser generates huge ultimate delays, enough to let
stacks overflow.

The size of the resulting linear processes after applying the improvements
of this commit is reduced with a factor 50 and 150.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14997 21252fe5-60e8-0310-8ff0-c1ae1001594e