There were two problems causing rewriting of the lambda term in this bug report to go astray. In revision 14659 a range based for was introduced, but in the for construct the variable count was also increased. This was forgotten in the adapted code. This means that lambda terms with multiple bound variables were dealt with incorrectly.
Furthermore, when there were multiple bound variables in a lambda term, and one of these variables would conflict with another bound variable, then the variables would be replaced by potentially renamed bound variables in the reversed order. This may be a remnant of partially replacing bound variables by a list.

A test case has been added.

This resolves the problem reported in #1426. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15012 21252fe5-60e8-0310-8ff0-c1ae1001594e