Repaired two errors in libraries/data/include/mcrl2/data/substitutions/mutable_indexed_substitution.h.
The first error regarded delivering expressions by reference from a vector. This is dangerous because
the vector can be moved leading a dangling reference.

The second problem was the calculation of variables in the right hand sides of the assignments in 
the substitution. Only the variables of the last expression was calculated, whereas it ought to calculate
all the variables in all right hand sides.

As far as I know these errors did not lead to problems in the code. 




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15019 21252fe5-60e8-0310-8ff0-c1ae1001594e