In ticket #1428 it was reported that equations were generated of the shape:
f == g = forall x0,x0: Nat. f(x0, x0) == g(x0, x0)

with bound variables with the same name. This was due to an index not
being increased. This has been repaired and a test case to signal this
has been added. This error was present since revision 12816 when 
equality on function symbols was added. 

This closes ticket #1428. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15027 21252fe5-60e8-0310-8ff0-c1ae1001594e