In #1445 it was reported that hidden actions in an action list were not properly renamed to tau when converting a .aut file into another .aut file.
This was due to the fact that the hidden_action_map was not properly copied
account when converting one transition system into another. Moreover, 
when saving a transition system this hidden_action_map was not applied,
leaving the original action names in place. 

The problem also occurred with the .dot, .fsm and .lts format.

This resolves #1445. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15083 21252fe5-60e8-0310-8ff0-c1ae1001594e