Fixed two bugs in the Fourier-Motzkin implementation
The first bug would occur when applied to an expression with
one non-linear equation, such as if(0 < x, x-2, 0) < 0. In that
case, constructing a linear_inequality would fail. The original
expression should be returned, but the proper return statement
was missing. Then, the empty disjunction from line 307 was
returned.

The second bug could cause the creation of quantifiers with
an empty list of variables. This does not pass the sanity checks
enabled in debug mode.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15110 21252fe5-60e8-0310-8ff0-c1ae1001594e