Removed some files related to an experiment to make the toolset parallel by allowing to create aterms in parallel. The conclustion of this experiment is that it appears to work. However, the creation of aterms is already so efficient that adding two compare and swap operators and making the reference counters atomic is already enough to delay making terms with a factor 4 (without running anything in parallel). Reading of terms can be done safely in parallel, but as the major time in the toolset is spent on creating terms, it is not really fruitful to make the toolset parallel unless far more than four processors are available.
Furthermore, move assign and move construct operators have been added explicitly
in such a way that they translate back to the standard operators on aterms. 
As it stands move assign is more efficient than assign, as it can swap the arguments.
E.g.:

   t1 = t2

in a move assign swaps t1 and t2 without changing reference counts. 

Move construct is exactly the same as a normal construct. In order
to perform a move construct in 

   aterm t1(t2)

one needs to leave t2 in a valid state, i.e. a state onto which the destructor
can be called decrementing it reference count. An experiment by assigning the
nullptr to t2 and not decreasing the reference count in case a term is the nullptr
works, but is sligtly less efficient for the non thread safe implementation. 
Most likely, for the thread safe implementation it can avoid an atomic increment,
and in that case it would be worthwhile.

Expecially, move assign operators have been made "noexcept". The implementation of
standard containers use this "noexcept" indication for a more efficient implementation.
The problem is that move semantics requires that when an exception occurs, the 
containers must be in a valid state. The extra required bookkeeping for this can be
ommitted if it is known that construction and assignment of components is will not
throw exceptions. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15121 21252fe5-60e8-0310-8ff0-c1ae1001594e