Removed the file libraries/atermpp/include/mcrl2/atermpp/container_utility.h which contained helper function of the shape "make_vector" that would get a finite number of elements and transform this into a vector.
The use of make_vector(a,b,...) has been replaced by ABC_TYPE_list({a,b,...})
where ABC_TYPE_list typically is variable_list of expression_list.

Using make_vector first a vector is constructed and subsequently this
is transformed into an aterm_list. This is not very efficient, although
it appears that this function was not used at places where efficiency 
matters. 





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15143 21252fe5-60e8-0310-8ff0-c1ae1001594e