Removed the file libraries/atermpp/include/mcrl2/atermpp/detail/aterm_list_utility.h.
In this file three overloads for + were defined. Two allowed
to add an element e to a list in the last or first position. Not only are
there standard functions to do this, the use of a symmetric symbol + for
this asymetric operation is a bad style.

The second overload was applicable to two lists of different type.
In l1 + l2 if the type of the elements of l1 was implicitly convertible
to elements of type l2, then this was a list of type l2.

The standard + on lists has been generalized such that it allows lists
with convertible elements and delivers the type of the most general list.
This makes the third operator superfluous.

A test has been added and all tests pass. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15150 21252fe5-60e8-0310-8ff0-c1ae1001594e