The previous commit message was completely mixed up, due to being in a hurry. Therefore, I sent its essence again. ---------------------------------------------------------------------
The failures inclusion algorithm turned out to be slow. This was
due to the fact that tuples of the form <imp_state, spec_states> were
put on the todo stack if they would not occur in the antichain set.
This completely in accordance with the abstractly described algorithm
from the literature on which the implementation is based.

All state pairs in the todo stack were only put in the antichain set when
they were investigated. The proper way to deal with this is to put
the states in the antichain set simultaneously when putting them on
the todo stack.

In concrete examples speedups of a factor 100 are achievable by this
change.

Problem was pointed out by Verum, as they were a little disappointed by
the performance.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15175 21252fe5-60e8-0310-8ff0-c1ae1001594e