Cleaned up the hash functions used in the aterm library. All hash functions are now only defined once.
Do not apply multiplication with a large prime number to
avoid dispersion of terms through memory.

Adapted the hash function for aterm_int's such that
they were not grouped with 8 different values in the
same hash bucket.

The code has become a few percent faster due to this change, 
although this may heavily depend on actual usage. 





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15328 21252fe5-60e8-0310-8ff0-c1ae1001594e