Maurice Laveaux observed that the frequently used function get_argument_of_higher_order_term returned a data_expression and not a const data_expression&. This leads to frequent increments and decrements of reference counts.
This has now been adapted.

Somewhat amazingly, this change appears to have a 5 to 10%
impact on performance, although the measurements for the compiling
rewriter are very unstable. Using the --timings flag, the new
routine requires 10s to 11.5s to generate the state space for 1394-fin, 
whereas the code with the old routine appears to require from 9.7s to 16.6s. 
I do not understand where these substantial differences come from. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15364 21252fe5-60e8-0310-8ff0-c1ae1001594e