Tim Willemse indicated that lpsxsim did not instantiate global variables.
Investigation taught that this was not actually true. The global variables
of a specification were instantiated, but only after the simulation framework
was set up. The net effect was unfortunately the same.

The problem should have been repaired now. Unfortunately, I cannot see why
starting the simulation framework was done so early, and moving it to a later
place in the code may have undesired side effects.

Furthermore, adapted the code to follow our own coding standard at some places. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@15385 21252fe5-60e8-0310-8ff0-c1ae1001594e