glue code for truly branching bisimulation
After discussion with JFG, we decided to add the glue code that allows
users to call the branching bisimulation algorithm that runs directly
on LTSs (instead of translating them to Kripke structures).  JFG will
hide the options from users. The goal is to run tests on this algorithm,
evaluate it and later publish it (if successful).