The result of lpsactionrename was sometimes quadratically too complex.
If in lpsactionrename a rename file was used of the shape

rename
   a(1) => a1;
   a(2) => a2;
   a(3) => a3;
   ...

then the resulting summands when renaming a summand with action a(t)
would look like

   t==1  ->  a1...  +
   t==2 && t!=1 -> a2... +
   t==3 && t!=1 && t!=2 -> a3... +

Such conditions can get quadratically large, while they
are equivalent to

   t==1  ->  a1...  +
   t==2 && 2!=1 -> a2... +
   t==3 && 3!=1 && 3!=2 -> a3... +

which after rewriting is only linearly large. The last conditions are
now being generated.

This was detected by Dennis Hendriks (TNO, ASML) who generated renaming
files of 160 renamings as give above. This would transform a linear
process which could generate its state space in seconds to one that
required an hour to be generated with the old quadratic translation.