Repaired the destruction of inequality caches.
The inequality caches were incorrectly destructed by only using delete,
and not using the destructor on the deleted fields. This has been
repaired by adding the destructor explicitly.

This is an instance where a std::unique_ptr could have been used,
except that a std::unique_pointer uses one word of memory too many.
It is unclear to me why this has been added in this way which appears
clumsy to me.

Hopefully this resolves the problem of Maurice where he detected
memory that was not properly deleted at the end of a program.