Removed the transformation strategy option from pbessolve.
The corresponding optimizations have been replaced by similar ones that operate on structure graphs. The new optimizations also work in combination with counter example generation.

Removed the equation map from pbessolve. Due to this the reset procedure has been removed as well, since it used this attribute. A new version of the reset procedure may have to be added later.