Made high way search truely random.
In highway search the number of states to be explored at each
level is limited. This allows to effectively peek deep inside
a huge statespace, without generating this state space. The
states that survive at each level are randomly chosen using the
rnd() function. Unfortunately, the seed was not set, meaning that
in each run exactly the same states were chosen. The seed is now
set depending on the time of day.