The translation of state labels of an .lts is particularly slow.
The code to translate state labels, which are balanced trees of data
expressions, into a string was very slow, due to an inefficient way
to walk through these state labels. The code now uses iterators and
the time to walk through these state labels has been completely
vaporised. This problem is only noticable when translating large .lts's,
which, as this appears to be mainly used in graphical tools, was
apparently not done too often.

The bottleneck is now the pp function on data expressions.