lpsuntime: put application of Fourier Motzkin behind an option
For some LPSs, lpsuntime would crash because of an exception in the
Fourier Motzkin algorithm. Since FM makes the working of the untime
algorithm unclear, it is only applied when asked so via an option.
Sometimes FM can remove a lot of sumvars, so I decided not to remove the
possibility from lpsuntime until it has been implemented in another
tool.
Fixes #1428