Added a warning for projection functions.
Added a warning when projection functions with the same name but with
different types in a single struct are being used. A typical examples is
sort S = struct s1(i:Int) | s2(i:Bool);
are begin used. This generally leads to type checking issues and is
hardly ever useful, as the functions i can easily be given different
names.