[mcrl2ide] added the option zoom code editors
This is added as an option in the view menu of mainwindow and the context menu of texteditor and has keyboard shorcuts.
Also changed most keyboard shortcuts to Qt default shortcuts and added shortcut for save as.
Also changed syntax highlighting such that it no longer has bold text, because mixing bold with not bold text messed up the monospacedness.