Made lpsrealzone a deprecated tool
lpsrealzone was an experimental tool that applies a transformation on LPSs.
The idea was to apply a zone-based abstraction on the LPSs, so a finite state space can be generated for them.
However, the approach was not very scalable and it place many restrictions on the input format, making the tool unusable in practice.
Furthermore, the code quality is very low, so making changes is very involved.
Therefore, I decided that this is a failed experiment, so it should be removed from the toolset at some point.
For now, I want to keep this around as a deprecated tool, in case I want to write about it in my thesis.