Reverted the changes to linearise.cpp as removing capture avoiding substitutions.
Capture avoiding substitutions on processes do not behave the samen
as the function substitute_pCRL expressions. Therefore, I reverted
these latter functions back. The difference shows itself on
lift3_init.mcrl2 and lift3_final.mcrl2, and most likely on many other
practical cases. The expectation is that this is due to handling
substitutions of assignments, but the precise reason is not yet clear.