Reverted the change to the native menu bar
Reverted the change to the native menu bar for the mac which
works on qt10 and higher, as we cannot get Qt11 to work on
the mac server on which TeamCity is running. If this works,
the ugly workaroun in qt_tool.h can be removed. Hopefully,
before the release.