Adapted a warning in mcrl22lps.
The tool mcrl22lps would issue warnings of the form:

Warning. Projection function  occurs multiple times with different sorts in struct empty | singleton(Xnat) | union(Xset, Xset).

as it detected that there were empty projection functions of sort Xnat
and Xset. But this is not a problem at all. Added a small check to
prevent the warning in this case. Reported by Maurice Laveaux.