Increased the version of QT for mac and set the menu bar to native again.
After discussion with Thomas, the least ugly solution that we could come
up with regarding the problem that versions of Qt lower than version
5.10 do not properly support menu bars on the Mac is to increase the
Qt version to be at least 5.10. This means that the look and feel of a
mac application is again exactly that what a mac user expects.