Added a new much faster substitution: replace_variables_capture_avoiding_with_an_identifier_generator
The functions replace_variables_capture_avoiding apply a substitution
while avoiding undesired capturing of variables by binders such as sum,
forall, exists, whr-clauses, lambda's. This function construct a set of
variables for every application of sigma, making this application slow.

In this commit a new additional substitution function is created with
the rather long name:

replace_variables_capture_avoiding_with_an_identifier_generator

This substitution requires an identifier generator, and it requires that
the substitution sigma can tell whether a variable occurs one of its
right hand sides. It does not need to construct an explicit set.

A substitution can be transformed to a substitution that maintains the
variables in its right hand side by a new class wrapper
maintain_variables_in_rh.

The new function
replace_variables_capture_avoiding_with_an_identifier_generator is made
available for all classes for which replace_variables_capture_avoiding
was also available.

The lineariser has been adapted to use the new substitution. For an
example such as merry_go_square actual linearisation is 10 times faster.
Furthermore, much of the code to calculate explicit right hand sides can
be removed also. This still needs to be done.

The latex file substitutions.tex is adapted to also contain the
definition of the new substitution.

Some points must still be addressed:
1) the name of the new function is too long. It should probably be shortened.
2) there are no tests. This is related to the problem that tests do not
   appear to work. See issue #1471.
3) Some of the code of replace_variables_capture_avoiding_with_an_identifier_generator
   is a direct copy of the code of replace_variables_capture_avoiding.
   It might be better to avoid this code copying, but as it stands, I do
   not see how.
4) The lineariser has unnecessary code to generate the variables in
   right hand sides. This must be cleaned up at some point.

This commit should be merged with the main branch after having been
reviewed by Wieger. It should be done after the August 2018 release as
the changes are substantial, especially to the lineariser.