Fixed many string comparisons in our CMake configuration.
 * This fixes the issue with library tests.

In CMake "if" conditions are special such that they dereference
variables without using brackets. For string comparisons this means
that if one side ends up naming a variable with a value assigned to
it then the value will be used for the comparison. As such to
avoid this issue we should use an explicit dereference, i.e. ${variable},
and use quotation marks to avoid further dereferencing.