Thorough rewrite of the branching bisimulation algorithm
This algorithm calculates the branching bisimulation quotient without
translating the labelled transition system to a Kripke structure.  It is
inspired in part by our GJKW 2017 paper.  That paper would reduce a LTS
in time O(m log m) because the translation to a Kripke structure introduces
O(m) additional states.  The algorithm in the present files avoids the
translation and tries to achieve O(m log n) time complexity on a LTS.

In particular, it uses *bunches* of transitions to register w. r. t. which
sets of transitions the partition is already stable:  ``If a state s has
a (non-inert) transition in a bunch, then every bottom state in the
block of s has a non-inert transition in that bunch.''  In contrast to
earlier versions of the algorithm, only non-inert transitions are
distributed over bunches; when an inert transition becomes non-inert,
we now quickly create a new bunch containing that transition.

Transitions are stored in four files to allow all operations needed.
This requires 9 pointers or size_t integer values per transition, which
is a little more than in the GJKW implementation, but we expect that the
overall memory use is still lower because fewer states and transitions
need to be stored.

I also renamed the filename _tb to _dnj.