Made the application iterator input, and removed instances of application(..) in favour of down_cast
The constructor of sort application(f,begin,end,transformer) was using a
forward iterator. This meant that there was a risk that the transformer
was applied multiple times, which can be time consuming, and even
undesired. This constructor now always uses an input iterator. In
application(f,begin,end) the iterator type is now either input or
forward, depending on the type of the iterators begin and end.

Furthermore, at a number of places application(t) for a data_expression
t has been replaced by atermpp::down_cast<application>(t) which is more
efficient as it does not copy the term.