Changed the run.py script for open-field-tictactoe.
The run.py script is much more efficient when the
pbes quantifier-all rewriter is applied before solving the
question whether there exists a winning strategy in the
game. The game can only be solved when putting quite
a number of pieces on the board initially. There is as
it stands no way to let model checking solve the game
as a whole.