Made the parser for parameterised parity games more robust
The rewriter that yields parameterised parity games does not give any
guarantee on the order of the condition and the propositional variable
instantiation within one clause/summand. This does not really matter for
PBESs constructed with lps2pbes (since the condition always is on the
left-hand side and the propositional variable instantiation is on the
right-hand side), but it does manifest itself when performing random
tests.