[mcrl2ide] split parsing and adding/editing in addeditpropertydialog
Now there is a button Parse to parse the formula and a button Add/Edit to only add/edit the property. This makes it possible to add/edit a property while it is not a valid mcf formula. Also the console now mentions that a specification is invalid after parsing (if it is indeed invalid).