Fixed both the strategy check and setting of the strategy in pbessolve.
The strategy check in pbessolve tried to prune the structure graph. But some subtle cases related to edge removal were not handled correctly. Now it uses a reachability algorithm starting from the initial state instead.

Choose a strategy in the set U if possible. This step is not very clearly described in the literature, but it turns out that it is necessary. Due to the faulty strategy check this error was not detected in the random tests.