[lpsxsim] initializing the simulation now longer blocks showing the UI
Fixes #1479
To have some indication that lpsxsim is busy initialising the simulation (for instance with infinitely branching lps), the status bar shows a message. This message does not stay up during the whole initialisation/updating process and I can't find out why.
During initialisation of a simulation it is possible to open another simulation instead. This will free the processor usage of the first simulation, but I haven't been able to free the memory usage as well.