Added an obvious ommission in the code generation when there are many rewrite rules.
For rewrite rules that have many cases, the jittyc rewriter generates
code with too many nesting of brackets. The gcc compilers refuse to
compile this. Auxiliary functions are generated to circumvent this
problem. A recent change in the generated code was not correctly applied
to the generation of these generated auxiliary functions. This was
spotted by Wieger Wesselink.

These auxiliary functions are typically necessary when using
struct types with many elements. For these types equality functions
are generated that have many rewrite rules of the shape:

   t1 == t2 = false

This is typically translated into:

if (arg1==t1)
{
    if (arg2==t2) return false
    {  ....

    }
}

As there are many cases, there are many nested curly brackets.
This is solved by introducing an auxiliary function when the
nesting of brackets is getting too deep.